requirements.txt
anthropic>=0.40.0
twilio>=9.0.0
playwright>=1.45.0
httpx>=0.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
